working framework; missing audit functions, currency exchange functionality, dateshifting

+++ R/misc_functions.R
@@ -1,5 +1,136 @@
 
 
+truncate_accounts <- function(string, level=NA){
+    output <- string
+    output[is.na(output)] <- ''
+    for(i in 1:length(string)){
+        break_pos <- gregexpr(':', output[i])[[1]]
+        if(!is.na(break_pos[level]) & break_pos[level]!='-1') output[i] <- substr(output[i], 1, break_pos[level]-1) 
+    }
+    return(output)
+}
+
+prepare_reports <- function(account_depth=2){
+  
+  if(!file.exists('./csv/accounts.csv')) stop("accounts.csv has not been created; run summarize_accounts first")
+  if(!file.exists('./csv/journal.csv')) stop("journal.csv has not been created; run prepare_journal first")
+      
+  a <- read.csv('./csv/accounts.csv', stringsAsFactors=FALSE)
+  d <- read.csv('./csv/journal.csv', stringsAsFactors=FALSE)
+
+  d$account <- truncate_accounts(d$account, level=account_depth)
+  a$account <- truncate_accounts(a$account, level=account_depth)
+
+  dir_init("./reports", overwrite=FALSE)
+
+  # income expense
+
+  min_year <- as.numeric(substr(min(d$date), 1, 4))
+  max_year <- as.numeric(substr(max(d$date), 1, 4))
+
+  year_range <- min_year : max_year
+
+  target_accounts <- unique(a$account[grep("^expenses|^income", a$account)])
+
+  ## annual
+
+  out <- matrix(NA, nrow=length(target_accounts), ncol=length(year_range))
+
+  colnames(out) <- year_range
+  rownames(out) <- target_accounts
+
+  for(i in 1:length(target_accounts)){
+    for(j in 1:length(year_range)){
+      start_date <- paste0(year_range[j], "-01-01")
+      stop_date <- paste0(year_range[j], "-12-31")
+      out[i,j] <- sum_amounts(journal=d, start_date=start_date, stop_date=stop_date, 
+        accounts=target_accounts[i])
+    }
+  }
+  
+  write.csv(out, "./reports/income_expense_annual.csv", row.names=TRUE) 
+  # if it ever gets to the point i want to break apart by decade, revisit this code
+
+  ## months, broken up by years
+
+  for(i in 1:length(year_range)){
+    out <- matrix(NA, nrow=length(target_accounts), ncol=12)
+    colnames(out) <- c("January", "February", "March", "April", "May", 
+        "June", "July", "August", "September", "October", "November", "December")
+    rownames(out) <- target_accounts
+    for(month in 1:12){
+      first_day <- paste(year_range[i], sprintf("%02d", month), "01", sep="-")
+
+      if(month==12){
+        last_day <- paste(year_range[i], sprintf("%02d", month), "31", sep="-")
+      } else {
+        last_day <- as.Date(paste(year_range[i], sprintf("%02d", month+1), "01", sep="-"))-1
+      }
+
+      for(j in 1:length(target_accounts)){
+        out[j,month] <- sum_amounts(journal=d, start_date=first_day, stop_date=last_day, 
+          accounts=target_accounts[j])
+      }
+    }
+    my_name <- paste0("./reports/income_expenses_monthly_", year_range[i], ".csv")
+    write.csv(out, my_name, row.names=TRUE)
+  }
+
+
+  # account_balances
+
+  min_year <- as.numeric(substr(min(d$date), 1, 4))
+  max_year <- as.numeric(substr(max(d$date), 1, 4))
+
+  year_range <- min_year : max_year
+
+  target_accounts <- a$account[grep("^assets|^liabilities|^equity", a$account)]
+
+  ## annual
+
+  out <- matrix(NA, nrow=length(target_accounts), ncol=length(year_range))
+
+  colnames(out) <- year_range
+  rownames(out) <- target_accounts
+
+  for(i in 1:length(target_accounts)){
+    for(j in 1:length(year_range)){
+      stop_date <- paste0(year_range[j], "-12-31")
+      out[i,j] <- sum_amounts(journal=d, stop_date=stop_date, 
+        accounts=target_accounts[i])
+    }
+  }
+  
+  write.csv(out, "./reports/account_balances_annual.csv", row.names=TRUE) 
+  # if it ever gets to the point i want to break apart by decade, revisit this code
+
+  ## months, broken up by years
+
+  for(i in 1:length(year_range)){
+    out <- matrix(NA, nrow=length(target_accounts), ncol=12)
+    colnames(out) <- c("January", "February", "March", "April", "May", 
+        "June", "July", "August", "September", "October", "November", "December")
+    rownames(out) <- target_accounts
+    for(month in 1:12){
+
+      if(month==12){
+        last_day <- paste(year_range[i], sprintf("%02d", month), "31", sep="-")
+      } else {
+        last_day <- as.Date(paste(year_range[i], sprintf("%02d", month+1), "01", sep="-"))-1
+      }
+
+      for(j in 1:length(target_accounts)){
+        out[j,month] <- sum_amounts(journal=d, stop_date=last_day, 
+          accounts=target_accounts[j])
+      }
+    }
+    my_name <- paste0("./reports/account_balances_monthly_", year_range[i], ".csv")
+    write.csv(out, my_name, row.names=TRUE)
+  }
+
+}
+
+
 absorb_entries <- function(){
 
   d <- read.csv("./csv/general_ledger.csv", stringsAsFactors=FALSE)
@@ -7,14 +138,14 @@ absorb_entries <- function(){
   inputs <- list.files("./primary_sources", pattern="*.csv", full.names=TRUE)
 
   add <- data.frame(date=character(), 
-    amount=character(),  tag=character(), notes=character(), 
-    account=character(), currency=character(), checksum=character(), 
-    tid=character(), balance=character()) 
+  amount=character(),  tag=character(), notes=character(), 
+  account=character(), currency=character(), checksum=character(), 
+  tid=character(), balance=character()) 
 
   if(length(inputs)>0){
-    for(i in 1:length(inputs)){
-      add <- rbind(add, read.csv(inputs[i], stringsAsFactors=FALSE))
-    }
+  for(i in 1:length(inputs)){
+    add <- rbind(add, read.csv(inputs[i], stringsAsFactors=FALSE))
+  }
   }
 
   # detect if entries are not present in the general ledger, timestamp and amount!
@@ -28,12 +159,12 @@ absorb_entries <- function(){
 
   if(length(keep)>0){
 
-    d <- rbind(d, add[keep,])
+  d <- rbind(d, add[keep,])
 
-    o <- order(d$date)
-    d <- d[o,]
+  o <- order(d$date)
+  d <- d[o,]
 
-    write.csv(d, './csv/general_ledger.csv', row.names=FALSE)
+  write.csv(d, './csv/general_ledger.csv', row.names=FALSE)
 
   }
 
@@ -41,82 +172,59 @@ absorb_entries <- function(){
 
 
 # can be used for flows or balances depending on what you pass in
-sum_amounts <- function(start_date=NA, stop_date=NA, accounts=NA){
-  d <- read.csv('./csv/journal.csv', stringsAsFactors=FALSE)
+sum_amounts <- function(journal, start_date=NA, stop_date=NA, accounts=NA){
+  d <- journal
   if(is.na(start_date)) start_date <- min(d$date)
   if(is.na(stop_date)) stop_date <- max(d$date)
-  if(is.na(accounts)) accounts <- sort(unique(d$account))
+  if(all(is.na(accounts))) accounts <- sort(unique(d$account))
   tar <- which(d$date >= start_date & d$date <= stop_date & d$account %in% accounts)
-  return(sum(d$amount[tar]))
+  output <- 0
+  if(length(tar)>0) output <- sum(d$amount[tar])
+  output <- round(output, 2)
+  return(output)
 }
 
 
-# unnecessary with grep involved!
-# surely this is simpler using regexpr???
-colon_reducer <- function(string, level=NA){
-    output <- string
-    output[is.na(output)] <- ''
-    for(i in 1:length(string)){
-        break_pos <- gregexpr(':', output[i])[[1]]
-        if(!is.na(break_pos[level]) & break_pos[level]!='-1') output[i] <- substr(output[i], 1, break_pos[level]-1) 
-    }
-    return(output)
-}
 
 texttab <- function(input.matrix, alignment=NA, hlines=NA, caption="", scale=NA){
-    output <- character(nrow(input.matrix))
-    for(i in 1:nrow(input.matrix)){
-        add.amps <- paste(input.matrix[i,], collapse=" & ")
-        output[i] <- paste(add.amps, "\\\\", sep=" ")
-    }
-    if(all(!is.na(hlines))){
-        for(i in 1:length(hlines)) output <- append(output, "\\hline", hlines[i]+(i-1))
-    }
-    return(output)
+  output <- character(nrow(input.matrix))
+  for(i in 1:nrow(input.matrix)){
+    add.amps <- paste(input.matrix[i,], collapse=" & ")
+    output[i] <- paste(add.amps, "\\\\", sep=" ")
+  }
+  if(all(!is.na(hlines))){
+    for(i in 1:length(hlines)) output <- append(output, "\\hline", hlines[i]+(i-1))
+  }
+  return(output)
 }
 
-# deprecated IMO
-journal_tabulator <- function( data, account_target, time_target, as.tex=FALSE, na.blank=TRUE ){
-    valid_rows <- which( ( data$y %in% time_target | data$ym %in% time_target ) & 
-            data$account1 %in% account_target )
-    data <- data[valid_rows,]
-    if( any( time_target %in% data$y ) ) x <- tapply( data$amount, list( data$account2, data$y ), sum )
-    if( any( time_target %in% data$ym ) ) x <- tapply( data$amount, list( data$account2, data$ym ), sum )
-    if( length(time_target) == 1 & time_target %in% data$ym ) x <- tapply( data$amount, list( data$account2, data$week ), sum )
-    if( nrow(x) == 0 ) stop('no entries found')
-    for( i in 1:ncol(x) ) x[,i] <- round(x[,i])
-    total <- apply( x, 2, function(z) sum(z, na.rm=TRUE) )
-    x <- rbind( x, total )
-    total <- apply( x, 1, function(z) sum(z, na.rm=TRUE) )
-    x <- cbind( x, total )
-    x <- cbind( rownames(x), x )
-    colnames(x)[1] <- 'account'
-    if(na.blank) x[is.na(x)] <- '.'
-    if(as.tex){
-        net_balance_tx <- x
-        net_balance_tx <- rbind( colnames(net_balance_tx), net_balance_tx )
-        net_balance_tx <- texttab(net_balance_tx, hline=c(1, nrow(net_balance_tx)-1, nrow(net_balance_tx)) )
-        net_balance_tx <- gsub( 'NA', '.', net_balance_tx )
-        x <- net_balance_tx
-    }
-    return(x)
+
+fix_dates <- function(dates){
+  if(length(grep('/', dates))>0){
+    last_four <- substr(dates, nchar(dates)-3, nchar(dates))
+    lower_y <- grep('/', last_four)
+    upper_y <- which(!1:nrow(d) %in% lower_y)
+    dates[upper_y] <- as.character(as.Date(dates, '%m/%d/%Y'))
+    dates[lower_y] <- as.character(as.Date(dates, '%m/%d/%y'))
+  }
+  dates <- as.Date(dates)
+  return(dates)
 }
 
-exchanger <- function(journal, prices, currency){
-    d <- journal
-    d$date <- as.Date(d$date)
-    prices$date <- as.Date(prices$date)
-    for(i in 1:nrow(d)){
-        if(d$currency[i]!=currency & d$currency[i] %in% prices$numerator){
-            currency_rows <- which(prices$numerator==d$currency[i])
-            currency_dates <- prices$date[currency_rows]
-            diffs <- abs(d$date[i] - currency_dates)
-            best_row <- currency_rows[which.min(diffs)]
-            d$amount[i] <- prices$price[best_row]*d$amount[i]
-            d$currency[i] <- paste0(currency,'-eqv')
-        }
-    } # ok good, but what if the numerare isn't in numerare commodity?
-    return(d)
+exchange <- function(journal, prices, currency){
+  journal$date <- fix_dates(journal$date)
+  prices$date <- fix_dates(prices$date)
+  for(i in 1:nrow(journal)){
+    if(journal$currency[i]!=currency & journal$currency[i] %in% prices$numerator){
+      currency_rows <- which(prices$numerator==journal$currency[i])
+      currency_dates <- prices$date[currency_rows]
+      diffs <- abs(journal$date[i] - currency_dates)
+      best_row <- currency_rows[which.min(diffs)]
+      journal$amount[i] <- prices$price[best_row]*journal$amount[i]
+      journal$currency[i] <- paste0(currency,'-eqv')
+    }
+  } # ok good, but what if the numerare isn't in numerare commodity?
+  return(journal)
 }
 
 
@@ -125,13 +233,7 @@ prepare_journal <- function(){
 
   d <- read.csv('./csv/general_ledger.csv', stringsAsFactors=TRUE)
 
-  if(length(grep('/', d$date))>0){
-      last_four <- substr(d$date, nchar(d$date)-3, nchar(d$date))
-      lower_y <- grep('/', last_four)
-      upper_y <- which(!1:nrow(d) %in% lower_y)
-      d$date[upper_y] <- as.character(as.Date(d$date, '%m/%d/%Y'))
-      d$date[lower_y] <- as.character(as.Date(d$date, '%m/%d/%y'))
-  }
+  d$date <- fix_dates(d$date)
   d$date <- as.character(d$date)
 
   drop <- which(d$tid=='')
@@ -158,7 +260,7 @@ prepare_journal <- function(){
 }
 
 
-balance_books <- function(){
+balance_accounts <- function(){
 
   d <- read.csv('./csv/general_ledger.csv', stringsAsFactors=FALSE)
   a <- read.csv('./csv/accounts.csv', stringsAsFactors=FALSE)
@@ -173,15 +275,15 @@ balance_books <- function(){
   
   for(i in 1:length(transaction_list)){
 
-    tid_rows <- which(d$tid==transaction_list[i])
+  tid_rows <- which(d$tid==transaction_list[i])
 
-    transaction_balance[i] <- sum(d$amount[tid_rows])
+  transaction_balance[i] <- sum(d$amount[tid_rows])
 
-    valid_tags[i] <- all(d$tag[tid_rows] %in% a$account)
-    # excludes multi-account tags, e.g. 'misc'
+  valid_tags[i] <- all(d$tag[tid_rows] %in% a$account)
+  # excludes multi-account tags, e.g. 'misc'
 
-    # have to convert into a single currency first to confirm balances
-    # if it doesn't balance, it might be b/c....
+  # have to convert into a single currency first to confirm balances
+  # if it doesn't balance, it might be b/c....
 
   }
 
@@ -189,15 +291,15 @@ balance_books <- function(){
 
   if(length(mirror_these)>0){
    
-    tar <- which(d$tid %in% mirror_these)
+  tar <- which(d$tid %in% mirror_these)
 
-    add <- d[tar,]
+  add <- d[tar,]
 
-    add$tag <- d$account[tar]
-    add$account <- d$tag[tar]
-    add$amount <- (-1)*add$amount
+  add$tag <- d$account[tar]
+  add$account <- d$tag[tar]
+  add$amount <- (-1)*add$amount
 
-    d <- rbind(d, add)
+  d <- rbind(d, add)
 
   }
   # order ledger by transaction dates
@@ -229,19 +331,19 @@ summarize_accounts <- function(){
 
   for(i in 1:length(account_list)){
 
-    involved <- which(d$account == account_list[i] | d$tag==account_list[i])
-    n_postings[i] <- sum(involved)
-    start_date[i] <- min(d$date[involved])
-    last_date[i] <- min(d$date[involved])
-    net_flow[i] <- sum(d$amount[involved])
-    n_checksums[i] <- sum(!is.na(d$checksum[d$account==account_list[i]]))
+  involved <- which(d$account == account_list[i] | d$tag==account_list[i])
+  n_postings[i] <- length(involved)
+  start_date[i] <- min(d$date[involved])
+  last_date[i] <- min(d$date[involved])
+  net_flow[i] <- sum(d$amount[involved])
+  n_checksums[i] <- sum(!is.na(d$checksum[d$account==account_list[i]]))
 
   }
 
   # also count how many other accounts these accounts share a tid with...hmmm
 
   output <- data.frame(account=account_list, n_postings, start_date, 
-    last_date, net_flow, n_checksums)
+  last_date, net_flow, n_checksums)
  
   write.csv(output, "./csv/accounts.csv", row.names=FALSE)
 
@@ -264,8 +366,8 @@ seperate_accounts <- function(){
 
   for(i in 1:length(account_names)){
 
-    tar <- which(d$account==a$account[i])
-    write.csv(d[tar,], account_names[i], row.names=FALSE)
+  tar <- which(d$account==a$account[i])
+  write.csv(d[tar,], account_names[i], row.names=FALSE)
 
   }
 
@@ -291,31 +393,40 @@ merge_accounts <- function(){
 
 
 id_maker <- function(n, reserved='', seed=NA, nchars=NA){
-    my_let <- letters 
-    my_num <- 0:9 
-    if(is.na(seed) | !is.numeric(seed)) set.seed(as.numeric(as.POSIXlt(Sys.time())))
-    if(!is.na(seed) & is.numeric(seed)) set.seed(seed)
-    output <- replicate(n, paste(sample(c(my_let, my_num), nchars, replace=TRUE), 
-        collapse=''))
+  my_let <- letters 
+  my_num <- 0:9 
+  if(is.na(seed) | !is.numeric(seed)) set.seed(as.numeric(as.POSIXlt(Sys.time())))
+  if(!is.na(seed) & is.numeric(seed)) set.seed(seed)
+  output <- replicate(n, paste(sample(c(my_let, my_num), nchars, replace=TRUE), 
+    collapse=''))
+  rejected <- duplicated(output) | output %in% reserved | substr(output, 1, 1) %in% my_num
+  while(any(rejected)){
+    output <- output[-which(rejected)]
+    remaining <- n-length(output)
+    output <- c(output, replicate(remaining, paste(sample(c(my_let, my_num), nchars, 
+      replace=TRUE), collapse='')))
     rejected <- duplicated(output) | output %in% reserved | substr(output, 1, 1) %in% my_num
-    while(any(rejected)){
-        output <- output[-which(rejected)]
-        remaining <- n-length(output)
-        output <- c(output, replicate(remaining, paste(sample(c(my_let, my_num), nchars, 
-            replace=TRUE), collapse='')))
-        rejected <- duplicated(output) | output %in% reserved | substr(output, 1, 1) %in% my_num
-    }
-    output
+  }
+  output
 }
 
-dir_init <- function(path, verbose=FALSE){
-    if(substr(path, 1, 2)!='./') stop('path argument must be formatted
-        with "./" at beginning')
-    contents <- dir(path, recursive=TRUE)
-    if(verbose){
+dir_init <- function(path, verbose=FALSE, overwrite=TRUE){
+  if(substr(path, 1, 2)!='./') stop('path argument must be formatted
+    with "./" at beginning')
+  contents <- dir(path, recursive=TRUE)
+  if(dir.exists(path)){
+    if(overwrite){
+      if(verbose){
         if(length(contents)==0) print(paste('folder ', path, ' created.', sep=""))
         if(length(contents)>0) print(paste('folder ', path, ' wiped of ', length(contents), ' files/folders.', sep=""))
+      }
+      if(dir.exists(path)) unlink(path, recursive=TRUE)
+      dir.create(path)
     }
-    if(dir.exists(path)) unlink(path, recursive=TRUE)
+  } else {
+    if(verbose){
+      print(paste('folder ', path, ' created.', sep=""))
+    }
     dir.create(path)
+  }
 }
